Servant.Session

Documentation

data SSession m k v

SSession m k v represents a session storage with keys of type k, values of type v, and operating under the monad m. The underlying implementation uses the 'wai-session' package, and any backend compatible with that package should work here too.